aegis-service/workspace/aegis_workspace/agents/travel_agent/requirements.txt
openai
python-dotenv
serpapi
requests
litellm
pydantic
typing-extensions
python-multipart
fastapi
uvicorn
tenacity
httpx